data_collection/requirements.txt
# python 3.9
selenium 
# download chromedriver from here https://googlechromelabs.github.io/chrome-for-testing/
# make sure to either place exe file in /usr/local/bin or some other directories in PATH 
# or create a link in /usr/local/bin by $ sudo ln -s /Users/hayley/Documents/chromedriver-mac-arm64/chromedriver /usr/local/bin/chromedriver
google-cloud-storage
# before then, download google cloud cli and set it up with your google cloud email address.
# how to install gcloud cli :  https://cloud.google.com/sdk/docs/install
# then initialize cli or do the following command: https://cloud.google.com/sdk/docs/initializing
# $gcloud init
# next, authorize gcloud cli or do the following : # https://cloud.google.com/docs/authentication/provide-credentials-adc#local-dev
# $gcloud auth login
# $gcloud auth application-default login 